Student Scores Vector manipulation
Create a class named Student  with the following specifications:
An instance variable named scores  to hold a student's  exam scores.
A void input() function that reads  integers and saves them to scores.
An int calculateTotalScore() function that returns the sum of the student's scores.

Shows some vector manipulations. Very Basic.